add ingress rules to allow TCP ports 22, 80, 443
We need to make TCP ports for certain services open for incoming
traffic, i.e. 22 for ssh, 80 for HTTP and 443 for HTTPS, because
22 is used for admin controls, 80 & 443 for web interfaces.

For that, we have to define resources for VPC, security groups,
an elastic IP, internet gateway, and routing tables. Without defining
those resources, the AWS EC2 instance cannot reachable from outside.