bountybox: read user_data directly instead of file provisioner
We don't need to make use of the file provisioner, because it's possible
to read user_data directly from the script.

This change results in a side-effect, that `$USER` is not any more
`ubuntu`, because it's executed in the early boot context. So we need to
use a plain string `ubuntu`.

Note, this mechanism works only for Ubuntu.