bountybox: run package installation at early-boot stage
Setting up basic packages should be done in the early-boot stage,
instead of a third-party script. That's because there's no way to
force strict orders to be applied between multiple third-party scripts.
So during `setup-docker.sh` installs packages, another script just
runs in parallel, which could fail because of missing packages.

Rename docker.yaml to cloudinit.yaml, and specify the list of packages
in the `packages` field in cloudinit.yaml. Also simplify downloading of
gpg keys.