Provision with remote-exec and bash
Allows us to avoid OS specific solutions like cloud-init or ignition and
with that redundant code, mostly. We can later resort to a more powerful
tool like mgmt that works across distros, if needed.

Run caddy and contained.af from docker containers for now to make the
setup as homogenous as possible.

Last but not least, do some terraform cleanup.