caddy: Add selinux label change attribute for volumemount
Caddy server has a host path volume mounted in the container, where this
container writes. With SELinux enabled this behavior won't be allowed
hence container on Fedora fails and never runs.

To modify the selinux label of the host file or directory being mounted
into the container, we have used selinux option `Z`.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>